added separate types file and convert some js files to ts
