+++ cga-frontend/src/composables/clipboard.ts
@@ -1,5 +1,5 @@
 export function useClipboard() {
-  const copyToClipboard = (text) => {
+  const copyToClipboard = (text: string): void => {
     navigator.clipboard.writeText(text);
   };
